Update: Coded login section of program.
The method loginSys(HashMap users) presents a standard username/password login form, then stores the matching user using the entered username as a key, finally, comparing the entered password to the match's username.

Also, changed delimiter in Scanner in initUsersMap() method to only use ':' as a delimiter as there was problems with reading the file of users.