Feature: Added registration system as method.
The method registerSys(HashMap  compare) takes a hashmap to add  a new user as the method is static.

A user registration form appears for entering user details and handling adding a new User to designated Users file.

Also has an optional consent input before providing a name for registering as an Agent if consent is given by the user.